install: send the real CI vendor name in the whoami user-agent
`bun pm whoami` built its `user-agent` header from a stub
`detect_ci_name` in `bun_install` that returned the literal `ci`
whenever the `CI` env var was present, with any value. So the header
ended in `ci/ci` for `CI=false` and `CI=0`, and never named the vendor
(`ci/github-actions`). `bun publish` used the real detector in
`bun_runtime::cli::ci_info`, which `bun_install` cannot depend on.

Move the detector and its vendor table to `bun_core::ci_info`. Delete
the install stub and the runtime copy. `whoami`, `publish`, and the
test runner (`test.only` and snapshot checks) now share one detector.
`CI=false` means not CI, and a vendor variable yields its name.
