+++ src/runtime/test_runner/ScopeFunctions.rs
@@ -481,7 +481,7 @@ impl ScopeFunctions {
 }
 
 fn error_in_ci(global: &JSGlobalObject, signature: &[u8]) -> JsResult<()> {
-    if crate::cli::ci_info::is_ci() {
+    if bun_core::ci_info::is_ci() {
         return Err(global.throw(format_args!(
             "{} is disabled in CI environments to prevent accidentally skipping tests. To override, set the environment variable CI=false.",
             bstr::BStr::new(signature)
